fix(debrid): address all 15 code review comments
- Fix 1: Add :debrid clause to History.get_adapter_module/1 so polling
  works for debrid clients

- Fix 2: test_download_client_connection now passes connection_settings
  (including provider key) for :debrid type, fixing provider config
  during UI connection tests

- Fix 3: Replace bare match in stream_all/4 with `with` so stream_each
  errors propagate instead of causing a MatchError crash

- Fix 4+5: Use actual .part file size as resume offset instead of
  download.bytes_pulled. Sync bytes_pulled before each file stream so
  crash-recovery and multi-file resume both use the correct offset

- Fix 6: remove_torrent now looks up and terminates any running Fetcher
  via a DB lookup by download_client_id, preventing orphaned fetchers

- Fix 7: Validate provider URLs via Shared.validate_download_url before
  persisting to Download.metadata in persist_r8_and_claim

- Fix 8: RealDebrid.list_jobs/2 acquires a RateLimiter slot per
  sub-request instead of relying on the single outer acquire, correctly
  protecting the 250/min budget during fan-out

- Fix 9: Add DNS resolution in private_host? to catch hostnames that
  resolve to RFC1918/link-local addresses (fail-open on DNS errors)

- Fix 10: Make RateLimiter.acquire atomic by routing through
  GenServer.call, eliminating the TOCTOU race between count and insert

- Fix 11: Pass prefetched_urls from persist_r8_and_claim to Fetcher.claim
  so the first run skips the redundant get_download_urls call; retries
  clear prefetched_urls and re-resolve from the provider

- Fix 12: Migration bytes_pulled column changed from :integer to :bigint

- Fix 13: Premiumize error bodies sanitized in wrap_pm/error_for_code
  before reaching Error structs; error logging in persist_r8_and_claim
  now sanitizes before inspect

- Fix 14: IPv6 ULA (fc00::/7) and link-local (fe80::/10) ranges added
  to SSRF blocklist; bracketed IPv6 literals stripped before parsing

- Fix 15: Implement retry logic using retries_left counter with
  exponential back-off (5s/10s/15s) before marking download failed

Agent-Logs-Url: https://github.com/getmydia/mydia/sessions/d6697126-cf55-476d-af0c-4446709c050d

+++ priv/repo/migrations/20260515135634_add_bytes_pulled_to_downloads.exs
@@ -3,7 +3,7 @@ defmodule Mydia.Repo.Migrations.AddBytesPulledToDownloads do
 
   def change do
     alter table(:downloads) do
-      add :bytes_pulled, :integer
+      add :bytes_pulled, :bigint
     end
   end
 end
